openhcl/tdx: hotfix for ap mailbox startup kernel requirements (#1562)
On TDX, the kernel uses the ACPI AP Mailbox protocol to start APs.
However, the kernel assumes that all kernel ram is identity mapped, as
the kernel will jump to a startup routine in any arbitrary kernel ram
range.

For now to hotfix main, describe 3GB of memory identity mapped in the
page table used by the mailbox assembly stub, so the kernel can start
APs regardless of how large the initial memory size was. An upcoming
change will instead have the bootshim modify the pagetable at runtime to
guarantee all ranges reported in the E820 map to kernel as ram are
mapped.

A future kernel change could remove this requirement entirely by making
the kernel spec compliant, and only require that the reset vector page
is identity mapped.