flowey: Properly propagate regen error codes (#1589)
When you sawn a Command in Rust exiting with a non-zero exit code is
still considered a "successful" execution, as you successfully waiting
for the process to finish running. You need to check its exit code to
know how the process really did. We weren't doing that in this shim,
resulting in `--check` errors being lost, which ultimately broke `xtask
fmt`.