Add OPENHCL_CONFIDENTIAL_DEBUG to manifest (#1627)
PR #1501 modified igvmfilegen to no longer explicitly add
OPENHCL_CONFIDENTIAL_DEBUG to the static command line if debug is
enabled. This broke SNP: SNP cannot yet directly read the debug bit, and
the vmbus relay depends on trusting the host for now. This change adds
OPENHCL_CONFIDENTIAL_DEBUG to the static command line explicitly in the
meantime (except for TDX, because it can directly read the debug bit).