scsi_buffers: return ReadOnly when attempting to write to read-only mapped buffers. (#1586)
Change an out of range error (for a zero-length paged range) into an
explicit ReadOnly error when the buffer is not writable. A non-writeable
buffer may occur, for example, when the guest indicates that a storage
packet is a WRITE (which should not require updating guest primary ram),
but storvsp still thinks it should write to guest memory.

This _should_ be a noop from an outside behavioral perspective. In
testing the usermode `storvsc` driver, we found that the error displayed
in this event is not helpful. An error is still correct. This should not
be a panic: the guest can control this behavior.

Suggested-By: @jstarks