vmbus_server: allow inspect to access ring buffers without full memory mapping (#1630)
This change updates how "inspect" accesses channel ring buffers by only
mapping the control page, and using `lockable_subrange` so that a
`GuestMemory` implementation that doesn't have a full mapping has the
opportunity to create a partial mapping.

This allows inspecting the ring buffers to work with alternative
`GuestMemory` implementations used in Hyper-V.

The data included in the "inspect" output is unchanged; the only change
is how it's retrieved.