Set can_trust_host if TDX debug bit is set (#1635)
Release backport of PR #1501 and PR #1627.

If the debug bit is set in the VM's TDX attributes, the host can be
trusted. This change gets the TD report in the boot shim and checks the
debug bit. If it's set, parse the dynamic command line to allow
enabling, e.g., confidential debugging.

Note that https://github.com/microsoft/OHCL-Linux-Kernel/pull/79 will
need to be resolved before hardware debug can be disabled in the
manifest.