vmgs: support ephemeral guest state in OpenHCL (#1612)
Support ephemeral guest state for everything that would otherwise use
the VMGS file in OpenHCL. This allows the host to avoid providing a VMGS
file over the GET entirely if it is not needed. This is useful for
testing purposes and could help to avoid issues with production
ephemeral VMs in the future (for example, boot failures caused by stale
guest state after a guest image swap).