uh_mem: Fix handling of unregistering multiply-registered overlay pages (#1633) (#1638)
When a GPN is registered as an overlay multiple times, we need to handle
unregistering it multiple times too. We don't see this in testing
because the only case we currently know of that registers the same GPN
multiple times never unregisters, but we should still be correct.
Multiple registration is a weird case anyways, so we just do our best.

Cherry-pick of #1633.