hcl: Private generalized get_vp_register functions (#1643)
The generalized functions here should not be used unless we know the
register is partition-wide, and not per-VP. This is easy to do
accidentally. Make these functions private so this is harder, and add
specialized accessors for some known good cases. Fix the bad case in the
tlb lock check.