flowey: add ADO trigger for tags (#1272)
This change introduces the ability to trigger pipelines when tags are
created. While developing this change I also noticed that the current
ADO CI triggers functionality did not generate correct yaml when an
empty vector of trigger branches is specified, so a fix for that issue
is included as well.

---------

Co-authored-by: Ben Hillis <benhill@ntdev.microsoft.com>