GET: Be tolerant of unexpected host responses (#1637) (#1641)
We have an ICM where we believe we are receiving a GET response intended
for a previous instance of the VM after a reset has completed. While we
are going to fix the host to not do this, we also need to tolerate it in
the meantime. We now warn on this happening, but then just drop the
unexpected response.

Clean cherry-pick of #1637