[disk_delay] A disk wrapper with runtime configurable I/O read and write delays to any underlying disk (#1614)
This change is part of changes for nvme keepalive testing and it will be
used in the nvme emulator to induce artificial delays and thus attempt
to capture perf of keepalive during servicing.
As a long term goal, the functionality in this crate will be expanded to
include configurable I/O drops, randomized slowdown and much more in
order to make testing of devices easier.

---------

Co-authored-by: Matt LaFayette (Kurjanowicz) <mattkur@microsoft.com>