firmware_uefi: EfiDiagnostics switch to polling instead of Arc<Mutex<>> (#1693)
This PR aims to take a different approach to flushing EfiDiagnostics on
Watchdog timeout.

- Removes Arc<Mutex<>>, which was originally used to share the
diagnostics service between UefiDevice and its WatchdogPlatform
- Creates a mesh::channel before creating the UefiDevice. The
watchdog_callback that gets added to the UefiDevice has its `on_timeout`
method modified to invoke the sender to send a notification.
- The mesh::receiver end gets sent down the UefiDevice, and it is used
in `poll_device()` to freely respond to watchdog events