Revert "virt_mshv_vtl: Wire up proper memory protections for the monitor page and enable it (#1541)" (#1731)
We've discovered that the current design for monitor pages on CVMs is
incompatible with OpenHCL as it relies on assigning to a shared page,
which we can't support. Remove all the code added for this, as it adds
complexity and is no longer needed. Vmbus will be going down a different
path in the future.